refactor(doctor): dedupe missing skill summaries
