+++ src/commands/doctor-skills.ts
@@ -16,32 +16,13 @@ import type { DoctorPrompter } from "./doctor-prompter.js";
 import {
   collectUnavailableAgentSkills,
   disableUnavailableSkillsInConfig,
+  formatMissingSkillSummary,
 } from "./doctor-skills-core.js";
 
 export {
   collectUnavailableAgentSkills,
   disableUnavailableSkillsInConfig,
 } from "./doctor-skills-core.js";
-
-function formatMissingSummary(skill: SkillStatusEntry): string {
-  const missing: string[] = [];
-  if (skill.missing.bins.length > 0) {
-    missing.push(`bins: ${skill.missing.bins.join(", ")}`);
-  }
-  if (skill.missing.anyBins.length > 0) {
-    missing.push(`any bins: ${skill.missing.anyBins.join(", ")}`);
-  }
-  if (skill.missing.env.length > 0) {
-    missing.push(`env: ${skill.missing.env.join(", ")}`);
-  }
-  if (skill.missing.config.length > 0) {
-    missing.push(`config: ${skill.missing.config.join(", ")}`);
-  }
-  if (skill.missing.os.length > 0) {
-    missing.push(`os: ${skill.missing.os.join(", ")}`);
-  }
-  return missing.join("; ") || "unknown requirement";
-}
 
 function formatInstallHints(skill: SkillStatusEntry): string[] {
   if (skill.install.length === 0) {
@@ -93,7 +74,7 @@ export function formatUnavailableSkillDoctorLines(skills: SkillStatusEntry[]): s
     "Some skills are allowed for this agent but are not usable in the current runtime environment.",
   ];
   for (const skill of skills) {
-    lines.push(`- ${skill.name}: ${formatMissingSummary(skill)}`);
+    lines.push(`- ${skill.name}: ${formatMissingSkillSummary(skill)}`);
     lines.push(...formatInstallHints(skill));
   }
   lines.push(`Disable unused skills: ${formatCliCommand("openclaw doctor --fix")}`);
